fix(api): paramless lock
Avoid locking API calls that do not contain any parameter such as `createGroup`.

When a meeting is created at BigBlueButton, the Meteor server sends an event requiring
the creation of a pad's group for the `notes` feature. Since `createGroup` does
not contains any parameters it's lock key was built using only the method's signature,
locking any other `createGroup` call within the server.

Included a timestamp for lock's key when the method lacks parameters.