build: convert package-lock.json to v2, add engines.node >= 16
bbb-pads was introduced in BBB 2.5 which has Node.js 16 (LTS) as the
baseline for most apps. That node version usually comes with npm 8 -
which converts lockfiles to v2 on install since it is the default.

This commit does two things: specify engines.node in package.json for
autodocumenting purposes and convert package-lock.json to
lockfileVersion 2 so that it isn't converted by the build system on
every build.